refactor(voip): extract ring group orchestrator
